If deleting specified images fails then try deleting all images instead.
This might be needed if we hit the 'image has dependent child images' issue.

+++ scripts/cleanImages.sh
@@ -10,5 +10,11 @@ if [ "$docker_images_list" == "" ]; then
     echo "No docker images on the agent"
 else
     echo "Clearing images: $docker_images_list"
-    docker rmi -f $docker_images_list
+    if docker rmi -f $docker_images_list ; then
+        echo "Deleting images was successful."
+    else
+        echo "Deleting specified images failed, so falling back to delete ALL images on system."
+        docker_images_list=$(docker images | awk '{print $3}' | uniq)
+        docker rmi -f $docker_images_list
+    fi
 fi
